Add files via upload
This is the current version of the variation. Now agents take their action based on the probability value generated. If the number generated is 1, the agent will have past action of his most immediate future opponent and therefore will make decision based on the most prominent action taken by its opponent. 

Thus, if the past history shows that the opponent cooperated more than defect then the agent will cooperate, else the agent will defect. The total payoff for each agent will be calculated and shown in  a leader board.

Our observation was that when players play their pure strategies, the defect-all strategy was the most dominant or is among the most dominant. However, it is the least dominant when this variation is experimented.